python/requirements.txt
avro==1.8.2
kafka-python==1.4.3
